requirements.txt
Django==1.6.5
Fabric==1.9.0
django-cors-headers==0.12
djangorestframework==2.3.14
dropbox==2.1.0
ecdsa==0.11
ipython==0.12.1
paramiko==1.14.0
pycrypto==2.6.1
readline==6.2.4.1
urllib3==1.9
wsgiref==0.1.2
